Relative dates and dependencies in append cmd
This commit adds support for relative dates and
dependencies in the append command. Relative dates
in due and start dates will be automatically converted
and dependenices will be recomputed.

When appending due/start dates, if the command already
has due or start dates, it will be replaced.
