+++ topydo/commands/AddCommand.py
@@ -22,14 +22,12 @@ from datetime import date
 from os.path import expanduser
 from sys import stdin
 
-from topydo.lib.Command import Command
+from topydo.lib.WriteCommand import WriteCommand
 from topydo.lib.Config import config
 from topydo.lib.prettyprinters.Numbers import PrettyPrinterNumbers
-from topydo.lib.RelativeDate import relative_date_to_date
-from topydo.lib.TodoListBase import InvalidTodoException
 
 
-class AddCommand(Command):
+class AddCommand(WriteCommand):
     def __init__(self, p_args, p_todolist, # pragma: no branch
                  p_out=lambda a: None,
                  p_err=lambda a: None,
@@ -70,69 +68,12 @@ class AddCommand(Command):
 
             return todo_text
 
-        def _postprocess_input_todo(p_todo):
-            """
-            Post-processes a parsed todo when adding it to the list.
-
-            * It converts relative dates to absolute ones.
-            * Automatically inserts a creation date if not present.
-            * Handles more user-friendly dependencies with before:, partof: and
-            after: tags
-            """
-            def convert_date(p_tag):
-                value = p_todo.tag_value(p_tag)
-
-                if value:
-                    dateobj = relative_date_to_date(value)
-                    if dateobj:
-                        p_todo.set_tag(p_tag, dateobj.isoformat())
-
-            def add_dependencies(p_tag):
-                for value in p_todo.tag_values(p_tag):
-                    try:
-                        dep = self.todolist.todo(value)
-
-                        if p_tag == 'after':
-                            self.todolist.add_dependency(p_todo, dep)
-                        elif p_tag == 'before' or p_tag == 'partof':
-                            self.todolist.add_dependency(dep, p_todo)
-                        elif p_tag.startswith('parent'):
-                            for parent in self.todolist.parents(dep):
-                                self.todolist.add_dependency(parent, p_todo)
-                        elif p_tag.startswith('child'):
-                            for child in self.todolist.children(dep):
-                                self.todolist.add_dependency(p_todo, child)
-                    except InvalidTodoException:
-                        pass
-
-                    p_todo.remove_tag(p_tag, value)
-
-            convert_date(config().tag_start())
-            convert_date(config().tag_due())
-
-            keywords = [
-                'after',
-                'before',
-                'child-of',
-                'childof',
-                'children-of',
-                'childrenof',
-                'parent-of',
-                'parentof',
-                'parents-of',
-                'parentsof',
-                'partof',
-            ]
-
-            for keyword in keywords:
-                add_dependencies(keyword)
-
-            if config().auto_creation_date():
-                p_todo.set_creation_date(date.today())
-
         todo_text = _preprocess_input_todo(p_todo_text)
         todo = self.todolist.add(todo_text)
-        _postprocess_input_todo(todo)
+        self.postprocess_input_todo(todo)
+
+        if config().auto_creation_date():
+            todo.set_creation_date(date.today())
 
         self.out(self.printer.print_todo(todo))
